Add serialize module
